+++ components/kyma-operator/pkg/kymaoperation/actions/step.go
@@ -1,4 +1,4 @@
-package kymainstallation
+package actions
 
 import (
 	"fmt"
