+++ components/kyma-operator/pkg/kymaoperation/actions/step_test.go
@@ -1,4 +1,4 @@
-package kymainstallation
+package actions
 
 import (
 	"errors"
