+++ components/kyma-operator/pkg/kymaoperation/actions/stepsfactory.go
@@ -1,4 +1,4 @@
-package kymainstallation
+package actions
 
 import (
 	"log"
