+++ components/kyma-operator/pkg/kymaoperation/actions/stepsfactory_test.go
@@ -1,4 +1,4 @@
-package kymainstallation
+package actions
 
 import (
 	"testing"
