--- components/kyma-operator/pkg/steps/doc.go
@@ -1,2 +1,0 @@
-//Package steps contains implementation for all installation steps. Not all are moved yet, work in progress
-package steps // import "github.com/kyma-project/kyma/components/kyma-operator/pkg/steps"
