--- components/kyma-operator/pkg/steps/installation.go
@@ -1,186 +1,0 @@
-package steps
-
-import (
-	"fmt"
-	"log"
-	"time"
-
-	"github.com/avast/retry-go"
-	"github.com/kyma-project/kyma/components/kyma-operator/pkg/actionmanager"
-	"github.com/kyma-project/kyma/components/kyma-operator/pkg/config"
-	internalerrors "github.com/kyma-project/kyma/components/kyma-operator/pkg/errors"
-	"github.com/kyma-project/kyma/components/kyma-operator/pkg/kymainstallation"
-	"github.com/kyma-project/kyma/components/kyma-operator/pkg/overrides"
-	serviceCatalog "github.com/kyma-project/kyma/components/kyma-operator/pkg/servicecatalog"
-	"github.com/kyma-project/kyma/components/kyma-operator/pkg/statusmanager"
-)
-
-type InstallStepsProvider interface {
-	For(installationData *config.InstallationData, overrideData overrides.OverrideData) (kymainstallation.InstallStepFactory, error)
-}
-
-type UninstallStepsProvider interface {
-	For(installationData *config.InstallationData) (kymainstallation.UninstallStepFactory, error)
-}
-
-//InstallationSteps .
-type InstallationSteps struct {
-	serviceCatalog        serviceCatalog.ClientInterface
-	errorHandlers         internalerrors.ErrorHandlersInterface
-	statusManager         statusmanager.StatusManager
-	actionManager         actionmanager.ActionManager
-	installStepsFactory   InstallStepsProvider
-	uninstallStepsFactory UninstallStepsProvider
-	backoffIntervals      []uint
-}
-
-// New .
-func New(serviceCatalog serviceCatalog.ClientInterface,
-	statusManager statusmanager.StatusManager, actionManager actionmanager.ActionManager,
-	stepFactoryCreator kymainstallation.StepFactoryCreator, backoffIntervals []uint) *InstallationSteps {
-	steps := &InstallationSteps{
-		serviceCatalog:        serviceCatalog,
-		errorHandlers:         &internalerrors.ErrorHandlers{},
-		statusManager:         statusManager,
-		actionManager:         actionManager,
-		installStepsFactory:   nil, //TODO: Implement
-		uninstallStepsFactory: nil, //TODO: Implement
-		backoffIntervals:      backoffIntervals,
-	}
-
-	return steps
-}
-
-//InstallKyma .
-func (steps *InstallationSteps) InstallKyma(installationData *config.InstallationData, overrideData overrides.OverrideData) error {
-
-	removeLabelAndReturn := steps.removeLabelOnError(installationData.Context.Name, installationData.Context.Namespace)
-
-	_ = steps.statusManager.InProgress("Verify installed components")
-
-	installStepsFactory, err := steps.installStepsFactory.For(installationData, overrideData)
-	if err != nil {
-		_ = steps.statusManager.Error("Kyma Operator", "Verifying installed components", err)
-		return err
-	}
-
-	stepList, err := installStepsFactory.InstallStepList()
-
-	if err != nil {
-		return removeLabelAndReturn(err)
-	}
-
-	err = steps.processComponents(removeLabelAndReturn, installationData.Action, stepList)
-	if err != nil {
-		return err
-	}
-
-	err = steps.statusManager.InstallDone(installationData.URL, installationData.KymaVersion)
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
-//UninstallKyma .
-func (steps *InstallationSteps) UninstallKyma(installationData *config.InstallationData) error {
-
-	removeLabelAndReturn := steps.removeLabelOnError(installationData.Context.Name, installationData.Context.Namespace)
-
-	err := steps.DeprovisionAzureResources(DefaultDeprovisionConfig(), installationData.Context)
-	steps.errorHandlers.LogError("An error during deprovisioning: ", err)
-
-	_ = steps.statusManager.InProgress("Verify components to uninstall")
-
-	uninstallStepsFactory, err := steps.uninstallStepsFactory.For(installationData)
-
-	if err != nil {
-		_ = steps.statusManager.Error("Kyma Operator", "Verify components to uninstall", err)
-		return err
-	}
-
-	stepList, err := uninstallStepsFactory.UninstallStepList()
-
-	if err != nil {
-		return removeLabelAndReturn(err)
-	}
-
-	err = steps.processComponents(removeLabelAndReturn, installationData.Action, stepList)
-	if steps.errorHandlers.CheckError("uninstall error: ", err) {
-		return err
-	}
-
-	err = steps.statusManager.UninstallDone()
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
-//PrintStep .
-func (steps *InstallationSteps) PrintStep(stepName string) {
-	log.Println("---------------------------")
-	log.Println(stepName)
-	log.Println("---------------------------")
-}
-
-func (steps *InstallationSteps) processComponents(removeLabelAndReturn func(err error) error, action string, stepList kymainstallation.StepList) error {
-
-	log.Println("Processing Kyma components")
-
-	logPrefix := action
-
-	for _, step := range stepList {
-
-		stepName := logPrefix + " component " + step.GetReleaseName()
-		_ = steps.statusManager.InProgress(stepName)
-
-		/*
-		   //We prepare entire list of steps upfront
-		   			step, err := stepsFactory.NewStep(component)
-		   			if err != nil {
-		   				return removeLabelAndReturn(err)
-		   			}
-		*/
-
-		steps.PrintStep(stepName)
-
-		err := retry.Do(
-			func() error {
-				processErr := step.Run()
-				if steps.errorHandlers.CheckError("Step error: ", processErr) {
-					_ = steps.statusManager.Error(step.GetReleaseName(), stepName, processErr)
-					return processErr
-				}
-				return nil
-			},
-			retry.Attempts(uint(len(steps.backoffIntervals))+1),
-			retry.DelayType(func(attempt uint, config *retry.Config) time.Duration {
-				log.Printf("Warning: Retry number %d (sleeping for %d[s]).\n", attempt+1, steps.backoffIntervals[attempt])
-				return time.Duration(steps.backoffIntervals[attempt]) * time.Second
-			}),
-		)
-
-		if err != nil {
-			return removeLabelAndReturn(fmt.Errorf("max number of retries reached during step: %s", stepName))
-		}
-
-		log.Println(stepName + "...DONE!")
-	}
-
-	log.Println(logPrefix + " Kyma components ...DONE!")
-
-	return removeLabelAndReturn(nil)
-}
-
-func (steps *InstallationSteps) removeLabelOnError(installationCrName, installationCrNamespace string) func(err error) error {
-	return func(err error) error {
-		removeLabelError := steps.actionManager.RemoveActionLabel(installationCrName, installationCrNamespace)
-		if steps.errorHandlers.CheckError("Error on removing label: ", removeLabelError) {
-			err = fmt.Errorf("%v; Error on removing label: %v", err, removeLabelError)
-		}
-		return err
-	}
-}
